# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PrivatePulse privacy-first analytics API
 * OpenAPI spec version: 0.2.0
 */

export interface AnalyticsSummary {
  totalEvents: number;
  uniquePages: number;
  /** @nullable */
  topEventName: string | null;
  changePercent: number;
  avgEventsPerDay: number;
}