# lib/api-zod/src/generated/types/apiKeyWithSecret.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PrivatePulse privacy-first analytics API
 * OpenAPI spec version: 0.2.0
 */

export interface ApiKeyWithSecret {
  id: string;
  workspaceId: string;
  name: string;
  prefix: string;
  createdAt: string;
  secret: string;
}