# lib/api-zod/src/generated/types/segment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PrivatePulse privacy-first analytics API
 * OpenAPI spec version: 0.2.0
 */
import type { SegmentCondition } from './segmentCondition';

export interface Segment {
  id: string;
  workspaceId: string;
  name: string;
  conditions: SegmentCondition[];
  createdAt: string;
}