# lib/api-zod/src/generated/types/segmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PrivatePulse privacy-first analytics API
 * OpenAPI spec version: 0.2.0
 */
import type { SegmentCondition } from './segmentCondition';

export interface SegmentInput {
  /** @minLength 1 */
  name: string;
  /** @minItems 1 */
  conditions: SegmentCondition[];
}