# lib/api-zod/src/generated/types/topReferrer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PrivatePulse privacy-first analytics API
 * OpenAPI spec version: 0.2.0
 */

export interface TopReferrer {
  referrer: string;
  count: number;
}